# backend/.build/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"GetUserResponse": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"phoneNumber": {
						"type": "string"
					},
					"cpf": {
						"type": "string"
					},
					"urlImage": {
						"type": "string"
					},
					"isPatient": {
						"type": "boolean"
					}
				},
				"required": [
					"id",
					"name",
					"email",
					"phoneNumber",
					"cpf",
					"urlImage",
					"isPatient"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateResponse": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserRequest": {
				"properties": {
					"name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"phoneNumber": {
						"type": "string"
					},
					"cpf": {
						"type": "string"
					},
					"urlImage": {
						"type": "string"
					},
					"isPatient": {
						"type": "boolean"
					}
				},
				"required": [
					"name",
					"email",
					"phoneNumber",
					"cpf",
					"urlImage",
					"isPatient"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GetReviewResponse": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"rating": {
						"type": "number",
						"format": "double"
					},
					"comment": {
						"type": "string"
					},
					"idCareProfessional": {
						"type": "number",
						"format": "double"
					},
					"idPatient": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"id",
					"rating",
					"comment",
					"idCareProfessional",
					"idPatient"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ReviewRequest": {
				"properties": {
					"rating": {
						"type": "number",
						"format": "double"
					},
					"comment": {
						"type": "string"
					},
					"idCareProfessional": {
						"type": "number",
						"format": "double"
					},
					"idPatient": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"rating",
					"comment",
					"idCareProfessional",
					"idPatient"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"GetAddressResponse": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"street": {
						"type": "string"
					},
					"number": {
						"type": "string"
					},
					"complement": {
						"type": "string"
					},
					"neighborhood": {
						"type": "string"
					},
					"city": {
						"type": "string"
					},
					"state": {
						"type": "string"
					},
					"postalCode": {
						"type": "string"
					},
					"country": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"street",
					"number",
					"neighborhood",
					"city",
					"state",
					"postalCode",
					"country"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AddressRequest": {
				"properties": {
					"street": {
						"type": "string"
					},
					"number": {
						"type": "string"
					},
					"complement": {
						"type": "string"
					},
					"neighborhood": {
						"type": "string"
					},
					"city": {
						"type": "string"
					},
					"state": {
						"type": "string"
					},
					"postalCode": {
						"type": "string"
					},
					"country": {
						"type": "string"
					}
				},
				"required": [
					"street",
					"number",
					"neighborhood",
					"city",
					"state",
					"postalCode",
					"country"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "backend",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
		"/users": {
			"get": {
				"operationId": "GetAllUsers",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/GetUserResponse"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreateUser",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CreateResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UserRequest"
							}
						}
					}
				}
			}
		},
		"/users/{id}": {
			"get": {
				"operationId": "GetUserById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GetUserResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateUser",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UserRequest"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteUser",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/reviews": {
			"get": {
				"operationId": "GetAll",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/GetReviewResponse"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Reviews"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "Create",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CreateResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Reviews"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ReviewRequest"
							}
						}
					}
				}
			}
		},
		"/reviews/{id}": {
			"get": {
				"operationId": "GetById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GetReviewResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Reviews"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "Update",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Reviews"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ReviewRequest"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "Remove",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Reviews"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/addresses": {
			"get": {
				"operationId": "GetAllAddresses",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/GetAddressResponse"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Addresses"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreateAddress",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CreateResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Addresses"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AddressRequest"
							}
						}
					}
				}
			}
		},
		"/addresses/{id}": {
			"get": {
				"operationId": "GetAddressById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GetAddressResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Addresses"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateAddress",
				"responses": {
					"204": {
						"description": "No Content"
					}
				},
				"tags": [
					"Addresses"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AddressRequest"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "RemoveAddress",
				"responses": {
					"204": {
						"description": "No Content"
					}
				},
				"tags": [
					"Addresses"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}